requirements.txt
absl-py==0.15.0
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.1
attrs==21.4.0
audioread==2.1.9
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.4
cryptography==3.4.8
decorator==5.1.1
Flask==2.0.3
flatbuffers==20210226132247
frozenlist==1.2.0
gast==0.4.0
google-auth==2.6.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.42.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
keras==2.6.0
Keras-Preprocessing==1.1.2
llvmlite==0.38.1
Markdown==3.3.4
MarkupSafe==2.0.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==5.1.0
numba==0.55.2
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pip==21.2.4
pooch==1.6.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.1.0
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyreadline==2.1
PySocks==1.7.1
requests==2.27.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
scikit-learn==1.1.1
scipy==1.7.3
setuptools==61.2.0
six==1.16.0
SoundFile==0.10.3.post1
tensorboard==2.6.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.6.0
tensorflow-cpu==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.1.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
wheel==0.35.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.13.3
yarl==1.6.3
zipp==3.8.0
